UENUM() enum class EJsutEnumTest : int32 {........} 这里做了限制,如果是uint8就不支持负数, 如果是int32就支持负数 就是int8和int32的范围而已

+++ Source/MyProject/SimpleTest.h
@@ -94,7 +94,7 @@ enum EEnumByteTest
 };
 
 UENUM()
-enum class EJsutEnumTest : uint8
+enum class EJsutEnumTest : int32
 {
 	EEnumTestBlue,
 	EEnumTestRed,
